requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-admin-interface==0.28.3
django-colorfield==0.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
python-slugify==8.0.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
